+++ src/index.html
@@ -23,7 +23,6 @@
   </footer>
 
   <script type="module" src="main.js"></script>
-  <script type="module" src="router.js"></script>
   <script src="https://www.gstatic.com/firebasejs/8.4.3/firebase-app.js"></script>
   <script src="https://www.gstatic.com/firebasejs/8.4.3/firebase-auth.js"></script>
   <script src="https://www.gstatic.com/firebasejs/8.4.3/firebase-firestore.js"></script>
